crypto-prices/price-fetcher/requirements.txt
numpy==1.24.3
pandas==2.0.3
python-binance==1.0.17